requirements.txt
transformers
opencv-python
pillow
torchvision
numpy
torch
gradio
captum
grad-cam